Question 1:
Finish matchDigits with the following requirements:

if x or y is negative, throw IllegalArgumentException
if x = 0 and y = 0, return 1
function should work if length of x or y is not the same ( x = 12, y = 132542 , returns 1)

Question 2:

Write a method multiplyEvens that returns the product of the first n even integers. For example:

Call	Output	Reason
multiplyEvens(1);	2	2 = 2
multiplyEvens(2);	8	2 * 4 = 8
multiplyEvens(3);	48	2 * 4 * 6 = 48
multiplyEvens(4);	384	2 * 4 * 6 * 8 = 384
You should throw an IllegalArgumentException if passed a value less than or equal to 0.